requirements.txt
numpy>=1.24
pandas>=2.0
pillow>=9.5
torch>=2.1
torchvision>=0.16
scikit-learn>=1.3
tqdm>=4.65
matplotlib>=3.7